Simplify transform system
